rudimentary jobs and associated skeleton tests
I have created these so there is something to build on next when
I start to code the JobFactory.  The tests and code in these classes
will be developed further once I understand how to mock or stub the
P6 api.  I don't want to use the real thing at all.