test for successful login
This is the second JMock test I have completed for the login to P6.

The TDD approach is proving to be a great learning experience as well
as a lot of work as I need to change several classes everytime I
realise I have to inject a dependency.

The lesson here is don't write a test to prove you can write a test
because every real test you write may force you to update those
skeleton tests for no value whatsoever.