Added test for TestConnectionJob failed login
A lot of files touched because I needed to change the abstract class Job and therefore all three of its sub classes.
Only the TestConnectionJob class is starting to get real code in it.
I realise the Job abstract class should be an interface but I am prepared to leave it as it is for now.
Remember atthis point I am learning JMock after little gaps.